+++ crates/oxc_linter/src/rules.rs
@@ -667,11 +667,16 @@ pub(crate) mod promise {
 }
 
 pub(crate) mod vitest {
+    pub mod consistent_each_for;
     pub mod consistent_test_filename;
     pub mod consistent_vitest_vi;
+    pub mod hoisted_apis_on_top;
     pub mod no_conditional_tests;
     pub mod no_import_node_test;
+    pub mod no_unneeded_async_expect_function;
+    pub mod prefer_called_once;
     pub mod prefer_called_times;
+    pub mod prefer_describe_function_title;
     pub mod prefer_to_be_falsy;
     pub mod prefer_to_be_object;
     pub mod prefer_to_be_truthy;
@@ -691,6 +696,7 @@ pub(crate) mod vue {
     pub mod define_props_declaration;
     pub mod define_props_destructuring;
     pub mod max_props;
+    pub mod no_arrow_functions_in_watch;
     pub mod no_deprecated_destroyed_lifecycle;
     pub mod no_export_in_script_setup;
     pub mod no_import_compiler_macros;
@@ -1333,11 +1339,16 @@ oxc_macros::declare_all_lint_rules! {
     unicorn::switch_case_braces,
     unicorn::text_encoding_identifier_case,
     unicorn::throw_new_error,
+    vitest::consistent_each_for,
     vitest::consistent_test_filename,
     vitest::consistent_vitest_vi,
+    vitest::hoisted_apis_on_top,
     vitest::no_conditional_tests,
     vitest::no_import_node_test,
+    vitest::no_unneeded_async_expect_function,
+    vitest::prefer_called_once,
     vitest::prefer_called_times,
+    vitest::prefer_describe_function_title,
     vitest::prefer_to_be_falsy,
     vitest::prefer_to_be_object,
     vitest::prefer_to_be_truthy,
@@ -1347,6 +1358,7 @@ oxc_macros::declare_all_lint_rules! {
     vue::define_props_declaration,
     vue::define_props_destructuring,
     vue::max_props,
+    vue::no_arrow_functions_in_watch,
     vue::no_deprecated_destroyed_lifecycle,
     vue::no_export_in_script_setup,
     vue::no_import_compiler_macros,